Reduce the number of initially needed packages for running the Installer.py
